오류 추가수정
