+++ src/index.ts
@@ -34,6 +34,13 @@ const port = process.env.PORT;
 /**
  * 공통 응답을 사용할 수 있는 헬퍼 함수 등록
  */
+declare module "express-serve-static-core" {
+  interface Response {
+    success: (data: any) => Response;
+    error: (params: { errorCode?: string; reason?: string | null; data?: any }) => Response;
+  }
+}
+
 app.use((req, res, next) => {
   res.success = (success) => {
     return res.json({ resultType: "SUCCESS", error: null, success });
